gallery and layout state modules
