+++ src/components/common/HeaderBar.vue
@@ -58,7 +58,7 @@ export default {
   },
   computed: {
     drawer_open () {
-      return this.$store.state.drawer_open
+      return this.$store.state.layout.drawer_open
     },
     has_add () {
       return (['GalleryPage', 'DeckPage'].indexOf(this.$route.name) > -1) && (this.$store.state.auth.username !== '')
