+++ src/components/deck/AddCard.vue
@@ -74,7 +74,7 @@ export default {
   },
   computed: {
     open_form () {
-      return this.$store.state.open_form
+      return this.$store.state.layout.open_form
     },
     deck_current () {
       return this.$store.state.deck_current
