+++ src/store.js
@@ -4,26 +4,21 @@ import api from '@/api/api'
 import { tools } from '@/utils/MStools'
 import { deckTools } from '@/utils/deckTools'
 import authStateModule from '@/state_modules/auth/authState'
+import galleryStateModule from '@/state_modules/gallery/galleryState'
+import layoutStateModule from '@/state_modules/layout/layoutState'
 
 Vue.use(Vuex)
-// Export store as function to ensure app settings are passed in before any initializations
+// Export store as function to ensure app settings
+//   are passed in before any initializations
 function builder (data) {
   return new Vuex.Store({
     modules: {
-      auth: authStateModule
+      auth: authStateModule,
+      gallery: galleryStateModule,
+      layout: layoutStateModule
     },
     state: {
-      // Sent to store during app initialization
       app_settings: data,
-
-      // Layout Data
-      drawer_open: false,
-      open_form: false,
-      form_tab: 'card',
-      page_loading: false,
-
-      // Gallery Data
-      gallery_list: [],
 
       // Decklist Data
       deck_lists: [],   // All decks from DB
@@ -38,53 +33,6 @@ function builder (data) {
     },
 
     mutations: {
-      // Layout Mutations
-      toggleDrawer (state, options = {}) {
-        state.drawer_open = options.hasOwnProperty('force')
-          ? options.force
-          : !state.drawer_open
-      },
-      toggleForm (state, options = {}) {
-        if (options.tab) {
-          state.form_tab = options.tab
-        }
-        state.open_form = !state.open_form
-      },
-      pageLoading (state, options) {
-        state.page_loading = options.loading
-      },
-
-      // Gallery Mutations
-      setGallery (state, options) {
-        const alters = options.alters
-        state.gallery_list = options.alters.sort(tools().sortBy('date', false))
-        state.gallery_list = state.gallery_list.map(deckTools().setAllCardsVisible)
-        this.commit('pageLoading', {loading: false})
-      },
-      addAlter (state, options) {
-        const alter = options.alter
-        state.gallery_list.push(alter)
-        state.gallery_list.sort(tools().sortBy('date', false))
-      },
-      sortGallery (state, options) {
-        state.gallery_list = state.gallery_list.sort(tools().sortBy(options.field, options.direction))
-      },
-      applyFilter (state, options) {
-        state.gallery_list = options.filter === 'All'
-          ? state.gallery_list.map(deckTools().setAllCardsVisible)
-          : state.gallery_list.map(card => {
-            card.visible = card.tags.includes(options.filter)
-            return card
-          })
-
-        // Apply Sort
-        state.gallery_list = state.gallery_list.sort(tools().sortBy(options.field, options.direction))
-      },
-      deleteAlter (state, options) {
-        const alterIdx = state.gallery_list.findIndex(alter => alter._id === options.id)
-        state.gallery_list.splice(alterIdx, 1)
-      },
-
       // Decklist Mutations
       setDecks (state, options) {
         state.deck_lists = options.decks
@@ -188,59 +136,6 @@ function builder (data) {
     },
 
     actions: {
-      // Gallery Actions
-      initGallery (state) {
-        state.commit('pageLoading', {loading: true})
-        api.get_cards()
-          .then(response => {
-            this.dispatch('combineGalleryListWithScryfall', {
-              'alters': response.data
-            })
-          })
-          .catch(err => {
-            console.warn('error getting altered card list: ')
-            console.error(err);
-            state.commit('pageLoading', {loading: false})
-          })
-      },
-      combineGalleryListWithScryfall (state, options) {
-        deckTools().combineScryfallData(options.alters)
-          .then(alters => {
-            state.commit('setGallery', {'alters': alters})
-          })
-      },
-      deleteAlter (state, options) {
-        api.delete_alter(options.id)
-          .then(() => {
-            state.commit('deleteAlter', options)
-          })
-          .catch(err => {
-            console.warn(' ** Err Deleting Alter')
-            console.error(err);
-          })
-      },
-      postAlter (state, options) {
-        api.post_gallery(options.alter)
-          .then(function(response) {
-            if (response.data.errors) {
-              console.warn(' ** Error posting new alter', response.data.message);
-            } else {
-              state.commit('toggleForm')
-              state.commit('addAlter', {'alter': options.alter})
-            }
-          })
-          .catch(err => console.warn(' ** error posting alter', error))
-      },
-      putAlter (state, options) {
-        api.update_gallery_card(options.alter)
-          .then(function(response) {
-            if (response.data.errors) {
-              console.warn(' ** Error updating alter', response.data.message);
-            }
-          })
-          .catch(error => console.error(' ** error updating alter', error))
-      },
-
       // Decklist Actions
       initDecks (state) {
         // store base list of existing altered cards
